% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab_bake.R
\name{crab_bake}
\alias{crab_bake}
\title{Executing CRAB Algorithm with crab_bake()}
\usage{
crab_bake(data_info)
}
\arguments{
\item{data_info}{A list containing the dataset, the original column names, and placeholder NULL
values for user-specified columns and parameters; the output from crab_prep()}
}
\value{
A data frame where each row corresponds to a tested k cluster. The data frame
includes the similarity score, runtime, and a best_clust column, which is blank
for all rows except the one with the lowest similarity score, where it is
marked as "BEST".
}
\description{
Retrieves the user-specified columns and parameters from the crab_prep()
object and uses them to execute crab().
}
\examples{
palmerpenguins::penguins |>
crab_prep() |>
crab_cols(c("flipper_length_mm", "bill_length_mm")) |>
crab_params(min_k = 2, max_k = 5, num_subsamples = 30, cluster_method = "kmeans") |>
crab_bake()
}
