% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab_cols.R
\name{crab_cols}
\alias{crab_cols}
\title{Choosing columns for dataset used in CRAB Algorithm}
\usage{
crab_cols(data_info, chosen_cols)
}
\arguments{
\item{data_info}{A list containing the dataset, the original column names, and placeholder
NULL values for user-specified columns and parameters; the output from crab_prep()}

\item{chosen_cols}{A list of column names from the dataset to be used in the analysis}
}
\value{
An updated list containing the dataset, its original column names, the
user-selected columns, and placeholder NULL values for the remaining parameters.
}
\description{
Updates crab_prep() object allowing the user to specify which dataset columns
should be used in crab().
}
\examples{
palmerpenguins::penguins |>
crab_prep() |>
crab_cols(c("flipper_length_mm", "bill_length_mm"))
}
