% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab_params.R
\name{crab_params}
\alias{crab_params}
\title{Choosing parameters for CRAB Algorithm}
\usage{
crab_params(
  data_info,
  min_k,
  max_k,
  num_subsamples,
  subsample_prop = 0.8,
  cluster_method,
  start_seed = 123
)
}
\arguments{
\item{data_info}{A list containing the dataset, the original column names, and placeholder
NULL values for user-specified columns and parameters; the output from crab_prep()}

\item{min_k}{The starting number of clusters (k) for evaluation}

\item{max_k}{The ending number of clusters (k) for evaluation}

\item{num_subsamples}{Number of resamples to generate}

\item{subsample_prop}{Proportion of the dataset to use for each resample. Defaults to 0.8.}

\item{cluster_method}{Specified clustering algorithm (i.e "kmeans")}

\item{start_seed}{Reproducibility seed for running crab()}
}
\value{
An updated list containing the dataset, its original column names, the
user-selected parameters, and placeholder NULL values for the specified columns.
}
\description{
Updates crab_prep() object allowing the user to specify which parameters should
be used in crab().
}
\examples{
palmerpenguins::penguins |>
crab_prep() |>
crab_params(min_k = 2, max_k = 5, num_subsamples = 10, cluster_method = "kmeans")
}
