% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab_prep.R
\name{crab_prep}
\alias{crab_prep}
\title{Initializing crab_setup with crab_prep}
\usage{
crab_prep(data)
}
\arguments{
\item{data}{The dataset to be analyzed}
}
\value{
A list containing the dataset, the original column names, and placeholder NULL
values for user-specified columns and parameters.
}
\description{
Acts as a structured container that holds the dataset, the original columns,
while reserving space for user-defined columns and parameters that will later
be used by crab_bake().
}
\examples{
palmerpenguins::penguins |> crab_prep()
}
