% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_matrix.R
\name{mean_matrix}
\alias{mean_matrix}
\title{Create Mean Matrix}
\usage{
mean_matrix(
  data = data,
  k = 3,
  subsample_prop = 0.8,
  num_subsamples = 10,
  start_seed = 123,
  cluster_method
)
}
\arguments{
\item{data}{The dataset to be analyzed}

\item{k}{The specific number of clusters (k) to evaluate. Defaults to 3.}

\item{subsample_prop}{Proportion of the dataset to use for each subsample. Defaults to 0.8.}

\item{num_subsamples}{Number of subsamples to generate. Defaults to 10.}

\item{start_seed}{Random seed for reproducibility. Defaults to 123.}

\item{cluster_method}{Specified clustering algorithm}
}
\value{
Produces a list of two elements. The first element, results, stores the
matrices generated by buddies_matrix(). The second element, score, represents
the final mean consistency matrix calculated from those results.
}
\description{
Computes a consistency score matrix for a specified k by measuring how
frequently each pair of observations is clustered together across subsamples.
Each element in the score matrix represents the sum of corresponding -1 and 1
values across subsamples, divided by the total number of valid comparisons.
This approach excludes zeros, which indicate observations not included in a
given resample. The function calls buddies_matrix() to obtain the matrices
for this calculation.
}
\examples{
data <- simulate_mvn(n = 10)
mean_matrix(data = data, k = 3, subsample_prop = 0.8, num_subsamples = 3,
            cluster_method = "kmeans")
}
