% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_contribution.R
\name{point_contribution}
\alias{point_contribution}
\title{Compute Contribution Score of Each Point in Relation to the CRAB Score.}
\usage{
point_contribution(data, mean_matrix)
}
\arguments{
\item{data}{The dataset to be analyzed}

\item{mean_matrix}{The output list produced by mean_matrix()}
}
\value{
Generates a data frame containing the index, individual point score, and
individual point contribution score for each data point with respect to the
CRAB score. Higher contribution values increase the CRAB score, while values
near zero have minimal impact.
}
\description{
Compute Contribution Score of Each Point in Relation to the CRAB Score.
}
\examples{
# data frame
mvn_df <- simulate_mvn(n = 400, variance = 0.4, desired_k = 3)

# mean matrix
mean_mat_mvn <- mean_matrix(data = mvn_df,
                            k = 3,
                            subsample_prop = 0.8,
                            num_subsamples = 10,
                            cluster_method = "kmeans")

# get point contributions
mvn_contributions <- point_contribution(data = mvn_df, mean_matrix = mean_mat_mvn)
head(mvn_contributions)

# visualize
mvn_df |>
ggplot(aes(x = X1, y = X2,
           color = mvn_contributions$singlePointContribution)) +
           scale_color_gradient(low = "darkblue", high = "red") +
           geom_point() +
           theme_minimal() +
           labs(title = "Original Data with Point Contribution",
                color = "Point Contribution")
}
