% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_simulations.R
\name{repeat_simulations}
\alias{repeat_simulations}
\title{Run CRAB Algorithm Using Simulated Data}
\usage{
repeat_simulations(
  crab_setup = NULL,
  num_runs,
  n = 100,
  variance,
  desired_k = 3,
  min_k,
  max_k,
  subsample_prop = 0.8,
  num_subsamples,
  start_seed = 123,
  cluster_method
)
}
\arguments{
\item{crab_setup}{A list containing a NULL dataset, NULL original column names, and values for
user-specified columns and parameters; the output from crab_prep()}

\item{num_runs}{Number of iterations}

\item{n}{Number of observations to generate. Defaults to 100. Note: If n does not split
evenly by desired_k, it will be rounded up.}

\item{variance}{List of variance values or single value used to construct the covariance
matrix for the dataset}

\item{desired_k}{The specific number of clusters (k) to evaluate. Defaults to 3.}

\item{min_k}{The starting number of clusters (k) for evaluation}

\item{max_k}{The starting number of clusters (k) for evaluation}

\item{subsample_prop}{Proportion of the dataset to use for each resample. Defaults to 0.8.}

\item{num_subsamples}{Number of resamples to generate}

\item{start_seed}{Random seed for reproducibility. Defaults to 123.}

\item{cluster_method}{Specified clustering algorithm}
}
\value{
A data frame where each row corresponds to a tested combination of k cluster,
covariance, and replication. The data frame includes the similarity score,
runtime, covariance, replication, and best_clust column. This column is blank
for all rows except the one with the lowest similarity score for that
combination within its own replication, where it is marked as "BEST".
}
\description{
Runs crab() repeatedly for a specified number of iterations, defined by num_runs.
}
\examples{
repeat_simulations(n = 100, num_runs = 5, variance = c(0.1, 0.2),
                   min_k = 2, max_k = 4, num_subsamples = 10,
                   cluster_method = "kmeans")
}
