% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_score.R
\name{similarity_score}
\alias{similarity_score}
\title{Obtain Similarity Score}
\usage{
similarity_score(mean_matrix = mean_mat)
}
\arguments{
\item{mean_matrix}{The output list produced by mean_matrix()}
}
\value{
Returns a single numeric value representing how closely the clustering results
align with the desired number of k clusters, based on the consistency of the
1s and -1s in the score matrix.
}
\description{
Computes a similarity score representing the squared distance from 1.
}
\examples{
data <- simulate_mvn(n = 10)
mean_matrix <- mean_matrix(data = data, k = 3, subsample_prop = 0.8,
num_subsamples = 3, cluster_method = "kmeans")
similarity_score(mean_matrix)
}
