% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mvn.R
\name{simulate_mvn}
\alias{simulate_mvn}
\title{Simulate Data from a Multivariate Normal Distribution}
\usage{
simulate_mvn(n = 100, variance = 0.4, desired_k = 3, start_seed = 123)
}
\arguments{
\item{n}{Number of observations to generate. Defaults to 100. Note: If n does
not split evenly by desired_k(), it will be rounded up.}

\item{variance}{Variance value used to construct the covariance matrix.
Defaults to 0.7}

\item{desired_k}{Desired number of clusters k. Defaults to 3.}

\item{start_seed}{Random seed for reproducibility. Defaults to 123.}
}
\value{
n x 2 standardized list with columns X1 and X2. Standardization is
performed within the function.
}
\description{
Generates a simulated dataset from a multivariate normal (MVN) distribution.
}
\examples{
simulate_mvn()
simulate_mvn(n = 150, covariance = 0.4)
simulate_mvn(desired_k = 4)
}
